services impl

+++ src/main/java/com/gwendal/magasinJPA/services/GenericService.java
@@ -5,8 +5,8 @@ import java.util.List;
 public interface GenericService<T> {
 	
 	public List<T> findAll();
-	public T findById(int id);
+	public T findById(Long id);
 	public T save(T entity);
-	public T update(int id, T entity);
-	public void delete(int id);
+	public T update(Long id, T entity);
+	public void delete(Long id);
 }
